JSON code will not be linted
Previously, JSHint used to lint JSON files. This was because
`self.view.settings().get('syntax')` only returns the directory of the
syntax's settings. In the case of the JSON syntax, we got
`Packages/JavaScript/JSON.tmLanguage`. Hence, your use of the regular
expression "JavaScript" on line 27 of JSHint.py returned true.

So I added an if-statement where if the syntax file's dirctory ends with
`JSON.tmLanguage`, `hasJsSyntax` will be set to false.

This tackles issue victorporof/Sublime-JSHint#29.